Update FriendPersistentInfo on 3DS
One of the DateTimes doesn't represent the date of the friendship, but
when the friend has updated their Mii.

I noticed this when playing with the friend servers researching
information about the data stored on the FRD sysmodule.

The friendship date is still stored internally, but it's not provided to
the user anymore.